fixed controller
